Factor out reactions test structure to utils file.
